move string concatenate to own header
