+++ src/system/popen.cxx
@@ -1,12 +1,12 @@
 #include "toolbox/system/popen.hpp"
-#include "toolbox/cpp/string.hpp"
+#include "toolbox/string/cat.hpp"
 
 namespace toolbox
 {
 
 Popen::Popen(std::string_view command, std::initializer_list<std::string_view> arguments, Mode mode)
 {
-    auto cmd = concatenate(command, arguments);
+    auto cmd = string::concatenate(command, arguments);
     fd = popen(cmd.c_str(), mode == Mode::Read ? "r" : "w");
 }
